Added enso.contrib.evaluate, which contains the 'evaluate' command.
This command is very similar to the 'calculate (four-function)' command in frozen Enso, only it's actually just a wrapper around Python's built-in eval() function.  It strips all builtins from the global namespace, so it should be hard/impossible (hopefully impossble?) to exploit it.  I saw this technique used somewere else--not sure where, but it's not my idea.

The big advantage here is that users have the full expressiveness of Python at their fingertips; not only can they do any kind of arithmetic, but they can also do things with strings, lists, dictionaries, et cetera.

The only major downside between this command and the standard calculate command is that this one doesn't use python Decimal objects to represent floating-point numbers, so weird things can happen in this regard (e.g., 0.1+0.1=0.2000000001 instead of 0.2).

The only other downside is that the errors raised by this command aren't as non-programmer-friendly as the old four-function calculate command, either.



git-svn-id: http://enso.googlecode.com/svn/trunk@26 8b7adc99-b347-0410-ae0a-d9e86c8d69b5